Map_Slider/requirements.txt
# CSC110 Fall 2020: Python libraries for final project

# Testing and code checking
pytest~=6.1.1
python-ta

# Graphics and data visualization
plotly~=4.11.0
pygame==2.0.0.dev10

# Map
geopandas~=0.8.1

# Reads file
pandas~=1.1.4

# Plots graphs
matplotlib~=3.3.3

# saves, resizes and loads images
Pillow~=8.0.1
